+++ packages/language/src/index.ts
@@ -1,6 +1,7 @@
 export * from './jpipe-module.js';
 export * from './jpipe-validator.js';
 export * from './jpipe-logger.js';
+export * from './jpipe-exclusions.js';
 export * from './generated/ast.js';
 export * from './generated/grammar.js';
 export * from './generated/module.js';